[nrf noup] cmake: scripts: add predefined configuration
This commit adds support for predefined configuration.  This is a
feature intended for deployed products consisting of multiple images.
It is crucial that any new versions of these images are linked to the
correct location. This is solved by this commit.

Also, this fixes the issue of building sub-images with mixed build
strategies.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>